# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ChemSafe Water Monitor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type AssessmentInputPipeMaterial = typeof AssessmentInputPipeMaterial[keyof typeof AssessmentInputPipeMaterial];


export const AssessmentInputPipeMaterial = {
  UPVC: 'UPVC',
  CPVC: 'CPVC',
  HDPE: 'HDPE',
  PVC: 'PVC',
  Unknown: 'Unknown',
} as const;

export type AssessmentInputGumType = typeof AssessmentInputGumType[keyof typeof AssessmentInputGumType];


export const AssessmentInputGumType = {
  solvent_cement: 'solvent_cement',
  rubber_ring: 'rubber_ring',
  unknown: 'unknown',
} as const;

export type AssessmentInputWaterFlowRate = typeof AssessmentInputWaterFlowRate[keyof typeof AssessmentInputWaterFlowRate];


export const AssessmentInputWaterFlowRate = {
  continuous: 'continuous',
  intermittent: 'intermittent',
  stagnant: 'stagnant',
} as const;

export type AssessmentInputBuriedOrExposed = typeof AssessmentInputBuriedOrExposed[keyof typeof AssessmentInputBuriedOrExposed];


export const AssessmentInputBuriedOrExposed = {
  buried: 'buried',
  exposed: 'exposed',
  mixed: 'mixed',
} as const;

export type AssessmentInputUseContext = typeof AssessmentInputUseContext[keyof typeof AssessmentInputUseContext];


export const AssessmentInputUseContext = {
  drinking: 'drinking',
  irrigation: 'irrigation',
  fish_pond: 'fish_pond',
  livestock: 'livestock',
} as const;

export type AssessmentInputSource = typeof AssessmentInputSource[keyof typeof AssessmentInputSource];


export const AssessmentInputSource = {
  interview: 'interview',
  form: 'form',
} as const;

/**
 * @nullable
 */
export type AssessmentInputConfidenceScores = {[key: string]: number} | null;

export interface AssessmentInput {
  pipe_material: AssessmentInputPipeMaterial;
  gum_type: AssessmentInputGumType;
  installation_age_years: number;
  /** @nullable */
  pipe_diameter_mm?: number | null;
  avg_temp_celsius: number;
  daily_uv_hours: number;
  water_flow_rate: AssessmentInputWaterFlowRate;
  buried_or_exposed: AssessmentInputBuriedOrExposed;
  use_context: AssessmentInputUseContext;
  /** @nullable */
  population_served?: number | null;
  source: AssessmentInputSource;
  /** @nullable */
  confidence_scores?: AssessmentInputConfidenceScores;
  /** @nullable */
  conversation_id?: number | null;
}

export type AssessmentRiskBand = typeof AssessmentRiskBand[keyof typeof AssessmentRiskBand];


export const AssessmentRiskBand = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
  critical: 'critical',
} as const;

/**
 * @nullable
 */
export type AssessmentConfidenceScores = {[key: string]: number} | null;

export interface Assessment {
  id: string;
  pipe_material: string;
  gum_type: string;
  installation_age_years: number;
  /** @nullable */
  pipe_diameter_mm?: number | null;
  avg_temp_celsius: number;
  daily_uv_hours: number;
  water_flow_rate: string;
  buried_or_exposed: string;
  use_context: string;
  /** @nullable */
  population_served?: number | null;
  source: string;
  risk_score: number;
  risk_band: AssessmentRiskBand;
  /** @nullable */
  score_min?: number | null;
  /** @nullable */
  score_max?: number | null;
  months_to_threshold: number;
  tdi: number;
  udi: number;
  /** @nullable */
  avg_confidence?: number | null;
  /** @nullable */
  confidence_scores?: AssessmentConfidenceScores;
  action_recommendation: string;
  /** @nullable */
  conversation_id?: number | null;
  created_at: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}