# lib/api-zod/src/generated/types/assessment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ChemSafe Water Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentConfidenceScores } from './assessmentConfidenceScores';
import type { AssessmentRiskBand } from './assessmentRiskBand';

export interface Assessment {
  id: string;
  pipe_material: string;
  gum_type: string;
  installation_age_years: number;
  /** @nullable */
  pipe_diameter_mm?: number | null;
  avg_temp_celsius: number;
  daily_uv_hours: number;
  water_flow_rate: string;
  buried_or_exposed: string;
  use_context: string;
  /** @nullable */
  population_served?: number | null;
  source: string;
  risk_score: number;
  risk_band: AssessmentRiskBand;
  /** @nullable */
  score_min?: number | null;
  /** @nullable */
  score_max?: number | null;
  months_to_threshold: number;
  tdi: number;
  udi: number;
  /** @nullable */
  avg_confidence?: number | null;
  /** @nullable */
  confidence_scores?: AssessmentConfidenceScores;
  action_recommendation: string;
  /** @nullable */
  conversation_id?: number | null;
  created_at: Date;
}