# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ChemSafe Water Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentInputBuriedOrExposed } from './assessmentInputBuriedOrExposed';
import type { AssessmentInputConfidenceScores } from './assessmentInputConfidenceScores';
import type { AssessmentInputGumType } from './assessmentInputGumType';
import type { AssessmentInputPipeMaterial } from './assessmentInputPipeMaterial';
import type { AssessmentInputSource } from './assessmentInputSource';
import type { AssessmentInputUseContext } from './assessmentInputUseContext';
import type { AssessmentInputWaterFlowRate } from './assessmentInputWaterFlowRate';

export interface AssessmentInput {
  pipe_material: AssessmentInputPipeMaterial;
  gum_type: AssessmentInputGumType;
  installation_age_years: number;
  /** @nullable */
  pipe_diameter_mm?: number | null;
  avg_temp_celsius: number;
  daily_uv_hours: number;
  water_flow_rate: AssessmentInputWaterFlowRate;
  buried_or_exposed: AssessmentInputBuriedOrExposed;
  use_context: AssessmentInputUseContext;
  /** @nullable */
  population_served?: number | null;
  source: AssessmentInputSource;
  /** @nullable */
  confidence_scores?: AssessmentInputConfidenceScores;
  /** @nullable */
  conversation_id?: number | null;
}